refactor(tls): split tls.rs (1911 lines) into tls/ module (P1-K #001)
The monolithic `src/tls.rs` mixed the error taxonomy, the rustls->TlsError
classifier, the config DTO, the CA-file loader, and the Pool builder in
one ~1900-line file with ~1100 lines of inline tests. Per the project's
file-size guideline (>=500 preferred, 1000 hard limit), split into a
directory module where each file has a single concern:

- src/tls/mod.rs       (43 lines) - //! doc + pub re-exports
- src/tls/error.rs    (487 lines) - TlsError enum + ctors + is_* helpers
- src/tls/classifier.rs (285 lines) - rustls::Error -> TlsError mapping
- src/tls/config.rs   (812 lines) - TlsConfig, TlsValidationMode, adapter
- src/tls/ca.rs       (124 lines) - CA PEM loading (re-exported as cert_utils)
- src/tls/pool.rs     (266 lines) - create_tls_connection + redact_url

The public API is preserved verbatim via `pub use` in `mod.rs`, so the
existing `gold_digger::tls::{TlsError, TlsConfig, TlsValidationMode,
create_tls_connection, redact_url, cert_utils, tls_config_from_url}`
import paths used by src/main.rs, src/cli.rs, src/exit.rs, and the
tests/ suite continue to resolve unchanged. Tests (inline #[cfg(test)])
moved to the submodule that owns the code under test; no test coverage
was lost.

No behavioural changes:
- All 45 inline tls:: unit tests pass.
- All external tls integration tests pass (tls_classifier_rcgen,
  tls_config_unit_tests, tls_backward_compatibility, tls_variants_test).
- Full `cargo nextest run` reports 552/552 passed, 13 skipped.
- `cargo clippy --all-targets -- -D warnings` is clean.
- `cargo deny check` reports advisories/bans/licenses/sources all ok.
